requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lyricsgenius==3.0.1
MarkupSafe==2.1.2
nltk==3.8.1
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
regex==2023.5.5
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
spotipy==2.23.0
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.2
vaderSentiment==3.3.2
Werkzeug==2.3.4
